Add contact page with map and a new font Clean up the commended code
