+++ src/modules/createPage.js
@@ -1,7 +1,8 @@
 import navBar from './navbar';
 import sidenav from './sidenav';
 import createMain from './main';
-import menu from './menu';
+import createMenu from './menu';
+import createContact from './contact';
 import footer from './footer';
 
 const insertAfter = (referenceNode, newNode) => {
@@ -21,4 +22,24 @@ const createPage = (mainHolder = main) => {
   insertAfter(content, sidenav);
 };
 
-export { createPage, insertAfter };
+const updatePage = event => {
+  event = event || window.event;
+  let target = event.target || event.srcElement,
+    text = target.textContent || target.innerText,
+    sections = ['main', 'our-menu', 'contactus'];
+
+  sections.forEach(s => {
+    if (document.getElementById(s) != null) document.getElementById(s).remove();
+  });
+
+  let m = document.getElementById('mobile-demo');
+  let choices = {
+    Home: () => insertAfter(m, createMain()),
+    Menu: () => insertAfter(m, createMenu()),
+    Contact: () => insertAfter(m, createContact())
+  };
+
+  choices[text]();
+};
+
+export { createPage, insertAfter, updatePage };
